bamboo/forms.py
from wtforms import Form, StringField, validators, PasswordField


class LoginForm(Form):
    username = StringField('Username', [validators.DataRequired()])
    password = PasswordField('Password', [validators.DataRequired()])
