requirements.txt
flask
flask-sqlalchemy
flask-admin
flask-login
flask-migrate
psycopg2
sqlalchemy-migrate
pandas
xlrd